requirements.txt
python-dotenv
flask-pymongo
flask
wtforms
flask-login
flask-wtf
email_validator